Improvements to jump system, fixes for null referece, and more
Fixed up the jump system so that the player can jump a little bit off a
platform, but can't control mid air. Fixed some empty referece
exceptions in 001Level2, and other misc improvements, incluiding the be
beginings of a particle system for when the hook hits an object, and
more.